performance improvements

+++ minifilter/snFilter/Communication.cpp
@@ -2,7 +2,8 @@
 
 #define POOL_FLAG_NON_PAGED 0x0000000000000040UI64 // Non paged pool NX
 
-NTSTATUS InitCommData() {
+NTSTATUS InitCommData()
+{
     HRESULT status;
     OBJECT_ATTRIBUTES oa;
     UNICODE_STRING uniString;
@@ -13,12 +14,13 @@ NTSTATUS InitCommData() {
     RtlInitUnicodeString(&uniString, ComPortName);
 
     status = FltBuildDefaultSecurityDescriptor(
-            &sd,
-            FLT_PORT_ALL_ACCESS); //  We secure the port so only ADMIN(s) & SYSTEM can access it.
+        &sd,
+        FLT_PORT_ALL_ACCESS); //  We secure the port so only ADMIN(s) & SYSTEM can access it.
     status = RtlSetDaclSecurityDescriptor(sd, TRUE, NULL,
                                           FALSE); // allow user application without admin to enter
 
-    if (NT_SUCCESS(status)) {
+    if (NT_SUCCESS(status))
+    {
         InitializeObjectAttributes(&oa, &uniString, OBJ_CASE_INSENSITIVE | OBJ_KERNEL_HANDLE, NULL, sd);
 
         status = FltCreateCommunicationPort(commHandle->Filter, &commHandle->ServerPort, &oa, NULL, RWFConnect,
@@ -34,19 +36,23 @@ NTSTATUS InitCommData() {
     return status;
 }
 
-BOOLEAN IsCommClosed() {
+BOOLEAN IsCommClosed()
+{
     return commHandle->CommClosed;
 }
 
-void CommClose() {
+void CommClose()
+{
     // FLT_ASSERT(IsCommClosed());
 
-    if (commHandle->ClientPort) {
+    if (commHandle->ClientPort)
+    {
         FltCloseClientPort(commHandle->Filter, &commHandle->ClientPort);
         commHandle->ClientPort = NULL;
     }
 
-    if (commHandle->ServerPort) {
+    if (commHandle->ServerPort)
+    {
         FltCloseCommunicationPort(commHandle->ServerPort);
         commHandle->ServerPort = NULL;
     }
@@ -59,35 +65,32 @@ RWFConnect(_In_ PFLT_PORT ClientPort, _In_opt_ PVOID ServerPortCookie,
            _In_reads_bytes_opt_(SizeOfContext) PVOID ConnectionContext, _In_ ULONG SizeOfContext,
            _Outptr_result_maybenull_ PVOID
 
-*ConnectionCookie)
+               *ConnectionCookie)
 {
-UNREFERENCED_PARAMETER(ServerPortCookie);
-UNREFERENCED_PARAMETER(ConnectionContext);
-UNREFERENCED_PARAMETER(SizeOfContext);
-UNREFERENCED_PARAMETER(ConnectionCookie = NULL
-);
+    UNREFERENCED_PARAMETER(ServerPortCookie);
+    UNREFERENCED_PARAMETER(ConnectionContext);
+    UNREFERENCED_PARAMETER(SizeOfContext);
+    UNREFERENCED_PARAMETER(ConnectionCookie = NULL);
 
-FLT_ASSERT(commHandle
-->ClientPort == NULL);
+    FLT_ASSERT(commHandle->ClientPort == NULL);
 
-//
-//  Set the user process and port. In a production filter it may
-//  be necessary to synchronize access to such fields with port
-//  lifetime. For instance, while filter manager will synchronize
-//  FltCloseClientPort with FltSendMessage's reading of the port
-//  handle, synchronizing access to the UserProcess would be up to
-//  the filter.
-//
+    //
+    //  Set the user process and port. In a production filter it may
+    //  be necessary to synchronize access to such fields with port
+    //  lifetime. For instance, while filter manager will synchronize
+    //  FltCloseClientPort with FltSendMessage's reading of the port
+    //  handle, synchronizing access to the UserProcess would be up to
+    //  the filter.
+    //
 
-commHandle->
-ClientPort = ClientPort;
-DbgPrint("!!! user connected, port=0x%p\n", ClientPort);
+    commHandle->ClientPort = ClientPort;
+    DbgPrint("!!! user connected, port=0x%p\n", ClientPort);
 
-return
-STATUS_SUCCESS;
+    return STATUS_SUCCESS;
 }
 
-VOID RWFDissconnect(_In_opt_ PVOID ConnectionCookie) {
+VOID RWFDissconnect(_In_opt_ PVOID ConnectionCookie)
+{
     UNREFERENCED_PARAMETER(ConnectionCookie);
 
     DbgPrint("!!! user disconnected, port=0x%p\n", commHandle->ClientPort);
@@ -108,7 +111,8 @@ VOID RWFDissconnect(_In_opt_ PVOID ConnectionCookie) {
 
 NTSTATUS
 RWFNewMessage(IN PVOID PortCookie, IN PVOID InputBuffer, IN ULONG InputBufferLength, OUT PVOID OutputBuffer,
-              IN ULONG OutputBufferLength, OUT PULONG ReturnOutputBufferLength) {
+              IN ULONG OutputBufferLength, OUT PULONG ReturnOutputBufferLength)
+{
     UNREFERENCED_PARAMETER(PortCookie);
     UNREFERENCED_PARAMETER(InputBufferLength);
 
@@ -118,49 +122,66 @@ RWFNewMessage(IN PVOID PortCookie, IN PVOID InputBuffer, IN ULONG InputBufferLen
     if (message == NULL)
         return STATUS_INTERNAL_ERROR; // failed message type
 
-    if (message->type == MESSAGE_ADD_SCAN_DIRECTORY) {
+    if (message->type == MESSAGE_ADD_SCAN_DIRECTORY)
+    {
         DbgPrint("Received add directory message\n");
         PDIRECTORY_ENTRY newEntry = new DIRECTORY_ENTRY();
-        if (newEntry == NULL) {
+        if (newEntry == NULL)
+        {
             return STATUS_INSUFFICIENT_RESOURCES;
         }
         NTSTATUS hr = CopyWString(newEntry->path, message->path, MAX_FILE_NAME_LENGTH);
-        if (!NT_SUCCESS(hr)) {
+        if (!NT_SUCCESS(hr))
+        {
             delete newEntry;
             return STATUS_INTERNAL_ERROR;
         }
         *ReturnOutputBufferLength = 1;
-        if (driverData->AddDirectoryEntry(newEntry)) {
-            *((PBOOLEAN) OutputBuffer) = TRUE;
+        if (driverData->AddDirectoryEntry(newEntry))
+        {
+            *((PBOOLEAN)OutputBuffer) = TRUE;
             DbgPrint("Added scan directory successfully\n");
             return STATUS_SUCCESS;
-        } else {
+        }
+        else
+        {
             delete newEntry;
-            *((PBOOLEAN) OutputBuffer) = FALSE;
+            *((PBOOLEAN)OutputBuffer) = FALSE;
             DbgPrint("Failed to add scan directory\n");
             return STATUS_SUCCESS;
         }
-    } else if (message->type == MESSAGE_REM_SCAN_DIRECTORY) {
+    }
+    else if (message->type == MESSAGE_REM_SCAN_DIRECTORY)
+    {
         PDIRECTORY_ENTRY ptr = driverData->RemDirectoryEntry(message->path);
         *ReturnOutputBufferLength = 1;
-        if (ptr == NULL) {
-            *((PBOOLEAN) OutputBuffer) = FALSE;
+        if (ptr == NULL)
+        {
+            *((PBOOLEAN)OutputBuffer) = FALSE;
             DbgPrint("Failed to remove directory\n");
             return STATUS_SUCCESS;
-        } else {
+        }
+        else
+        {
             delete ptr;
         }
-        *((PBOOLEAN) OutputBuffer) = TRUE;
+        *((PBOOLEAN)OutputBuffer) = TRUE;
         DbgPrint("Removed scan directory successfully\n");
         return STATUS_SUCCESS;
-    } else if (message->type == MESSAGE_GET_OPS) {
-        if (OutputBuffer == NULL || OutputBufferLength != MAX_COMM_BUFFER_SIZE) {
+    }
+    else if (message->type == MESSAGE_GET_OPS)
+    {
+        if (OutputBuffer == NULL || OutputBufferLength != MAX_COMM_BUFFER_SIZE)
+        {
             return STATUS_INVALID_PARAMETER;
         }
         driverData->DriverGetIrps(OutputBuffer, OutputBufferLength, ReturnOutputBufferLength);
         return STATUS_SUCCESS;
-    } else if (message->type == MESSAGE_SET_PID) {
-        if (message->pid != 0) {
+    }
+    else if (message->type == MESSAGE_SET_PID)
+    {
+        if (message->pid != 0)
+        {
             driverData->setPID(message->pid);
             driverData->setSystemRootPath(message->path);
             commHandle->CommClosed = FALSE;
@@ -169,9 +190,11 @@ RWFNewMessage(IN PVOID PortCookie, IN PVOID InputBuffer, IN ULONG InputBufferLen
         }
         return STATUS_INVALID_PARAMETER;
     }
-        // TODO: the kill code to gid
-    else if (message->type == MESSAGE_KILL_GID) {
-        if (OutputBuffer == NULL || OutputBufferLength != sizeof(LONG)) {
+    // TODO: the kill code to gid
+    else if (message->type == MESSAGE_KILL_GID)
+    {
+        if (OutputBuffer == NULL || OutputBufferLength != sizeof(LONG))
+        {
             return STATUS_INVALID_PARAMETER;
         }
         *ReturnOutputBufferLength = sizeof(LONG);
@@ -180,52 +203,57 @@ RWFNewMessage(IN PVOID PortCookie, IN PVOID InputBuffer, IN ULONG InputBufferLen
         ULONGLONG GID = message->gid;
         BOOLEAN isGidExist = FALSE;
         ULONGLONG gidSize = driverData->GetGidSize(GID, &isGidExist);
-        if (gidSize == 0 || isGidExist == FALSE) {
-            DbgPrint("!!! FS : Gid already ended or no such gid %d\n", GID);
-            *((PLONG) OutputBuffer) = STATUS_NO_SUCH_GROUP; // fail to kill process
+        if (gidSize == 0 || isGidExist == FALSE)
+        {
+            DbgPrint("!!! FS : Gid already ended or no such gid %d\n", (int)GID);
+            *((PLONG)OutputBuffer) = STATUS_NO_SUCH_GROUP; // fail to kill process
             return STATUS_SUCCESS;
         }
         // there is gid with processes
         PULONG
-                Buffer = (PULONG)
-        ExAllocatePool2(POOL_FLAG_NON_PAGED, sizeof(ULONG) * gidSize, 'RW');
-        if (Buffer == nullptr) {
+        Buffer = (PULONG)ExAllocatePool2(POOL_FLAG_NON_PAGED, sizeof(ULONG) * gidSize, 'RW');
+        if (Buffer == nullptr)
+        {
             DbgPrint("!!! FS : memory allocation error on non paged pool\n");
-            *((PLONG) OutputBuffer) = STATUS_MEMORY_NOT_ALLOCATED; // fail to kill process
+            *((PLONG)OutputBuffer) = STATUS_MEMORY_NOT_ALLOCATED; // fail to kill process
             return STATUS_SUCCESS;
         }
         ULONGLONG pidsReturned = 0;
         isGidExist = driverData->GetGidPids(GID, Buffer, gidSize, &pidsReturned);
-        if (isGidExist) { // got all irps and correct size
-            for (int i = 0; i < gidSize; i++) { // kill each process
+        if (isGidExist)
+        { // got all irps and correct size
+            for (int i = 0; i < gidSize; i++)
+            { // kill each process
                 CLIENT_ID clientId;
-                clientId.UniqueProcess = (HANDLE) Buffer[i];
+                clientId.UniqueProcess = (HANDLE)Buffer[i];
                 clientId.UniqueThread = 0;
 
                 OBJECT_ATTRIBUTES objAttribs;
                 NTSTATUS exitStatus = STATUS_FAIL_CHECK;
 
-                DbgPrint("!!! FS : Attempt to terminate pid: %d from gid: %d\n", Buffer[i], GID);
+                DbgPrint("!!! FS : Attempt to terminate pid: %d from gid: %d\n", Buffer[i], (int)GID);
 
                 InitializeObjectAttributes(&objAttribs, NULL, OBJ_KERNEL_HANDLE, NULL, NULL);
 
                 status = ZwOpenProcess(&processHandle, PROCESS_ALL_ACCESS, &objAttribs, &clientId);
 
-                if (!NT_SUCCESS(status)) {
-                    *((PLONG) OutputBuffer) = STATUS_FAIL_CHECK; // fail
+                if (!NT_SUCCESS(status))
+                {
+                    *((PLONG)OutputBuffer) = STATUS_FAIL_CHECK; // fail
                     DbgPrint("!!! FS : Failed to open process %d, reason: %d\n", Buffer[i], status);
                     continue; // try to kill others
                 }
                 status = ZwTerminateProcess(processHandle, exitStatus);
-                if (!NT_SUCCESS(status)) {
-                    *((PLONG) OutputBuffer) = STATUS_FAIL_CHECK; // fail
+                if (!NT_SUCCESS(status))
+                {
+                    *((PLONG)OutputBuffer) = STATUS_FAIL_CHECK; // fail
                     DbgPrint("!!! FS : Failed to kill process %d, reason: %d\n", Buffer[i], status);
                     status = NtClose(processHandle);
                     continue; // try to kill others
                 }
                 NtClose(processHandle);
 
-                DbgPrint("!!! FS : Termination of pid: %d from gid: %d succeeded\n", Buffer[i], GID);
+                DbgPrint("!!! FS : Termination of pid: %d from gid: %d succeeded\n", Buffer[i], (int)GID);
             }
         }
         ExFreePoolWithTag(Buffer, 'RW');
